Rename name of app because errmsg are much to long
