+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "com.sw_ss16.studyroompopulationpredicter"
+        applicationId "com.sw_ss16.lc_app"
         minSdkVersion 14
         targetSdkVersion 23
         versionCode 2
