+++ app/src/main/java/com/sw_ss16/lc_app/ui/ViewSamplesActivity.java
@@ -1,18 +1,16 @@
-package com.sw_ss16.studyroompopulationpredicter.ui;
+package com.sw_ss16.lc_app.ui;
 
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
 import android.support.v7.app.ActionBar;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 
-import butterknife.Bind;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
-import com.sw_ss16.studyroompopulationpredicter.R;
-import com.sw_ss16.studyroompopulationpredicter.ui.base.BaseActivity;
+import com.sw_ss16.lc_app.R;
+import com.sw_ss16.lc_app.ui.base.BaseActivity;
 
 /**
  * Activity demonstrates some GUI functionalities from the Android support library.
