+++ app/src/main/java/com/sw_ss16/lc_app/ui/base/BaseFragment.java
@@ -1,4 +1,4 @@
-package com.sw_ss16.studyroompopulationpredicter.ui.base;
+package com.sw_ss16.lc_app.ui.base;
 
 import android.app.Fragment;
 import android.view.LayoutInflater;
@@ -6,9 +6,9 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import butterknife.ButterKnife;
-import com.sw_ss16.studyroompopulationpredicter.util.LogUtil;
+import com.sw_ss16.lc_app.util.LogUtil;
 
-import static com.sw_ss16.studyroompopulationpredicter.util.LogUtil.makeLogTag;
+import static com.sw_ss16.lc_app.util.LogUtil.makeLogTag;
 
 /**
  * The base class for all fragment classes.
