+++ app/src/main/res/layout-sw600dp-land/include_list_view.xml
@@ -11,7 +11,7 @@
 
     <fragment
         android:id="@+id/article_list"
-        android:name="com.sw_ss16.studyroompopulationpredicter.ui.studyroom.StudyRoomListFragment"
+        android:name="com.sw_ss16.lc_app.ui.studyroom.StudyRoomListFragment"
         android:layout_width="0dp"
         android:layout_height="match_parent"
         android:layout_weight="3"
